# internal/api/api_v1/oas_router_gen.go
// Code generated by ogen, DO NOT EDIT.

package api_v1

import (
	"net/http"
	"net/url"
	"strings"

	"github.com/ogen-go/ogen/uri"
)

var (
	rn10AllowedHeaders = map[string]string{
		"POST": "Content-Type",
	}
	rn7AllowedHeaders = map[string]string{
		"PATCH": "Content-Type",
	}
	rn9AllowedHeaders = map[string]string{
		"POST": "Content-Type",
		"PUT":  "Content-Type",
	}
)

func (s *Server) cutPrefix(path string) (string, bool) {
	prefix := s.cfg.Prefix
	if prefix == "" {
		return path, true
	}
	if !strings.HasPrefix(path, prefix) {
		// Prefix doesn't match.
		return "", false
	}
	// Cut prefix from the path.
	return strings.TrimPrefix(path, prefix), true
}

// ServeHTTP serves http request as defined by OpenAPI v3 specification,
// calling handler that matches the path or returning not found error.
func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
	elem := r.URL.Path
	elemIsEscaped := false
	if rawPath := r.URL.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
			elemIsEscaped = strings.ContainsRune(elem, '%')
		}
	}

	elem, ok := s.cutPrefix(elem)
	if !ok || len(elem) == 0 {
		s.notFound(w, r)
		return
	}
	args := [4]string{}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/managers"

			if l := len("/managers"); len(elem) >= l && elem[0:l] == "/managers" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				switch r.Method {
				case "GET":
					s.handleManagersGetRequest([0]string{}, elemIsEscaped, w, r)
				default:
					s.notAllowed(w, r, notAllowedParams{
						allowedMethods: "GET",
						allowedHeaders: nil,
						acceptPost:     "",
						acceptPatch:    "",
					})
				}

				return
			}
			switch elem[0] {
			case '/': // Prefix: "/"

				if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "managerName"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case '/': // Prefix: "/tasks"

					if l := len("/tasks"); len(elem) >= l && elem[0:l] == "/tasks" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch r.Method {
						case "GET":
							s.handleTasksGetRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						case "POST":
							s.handleTaskPostRequest([1]string{
								args[0],
							}, elemIsEscaped, w, r)
						default:
							s.notAllowed(w, r, notAllowedParams{
								allowedMethods: "GET,POST",
								allowedHeaders: rn10AllowedHeaders,
								acceptPost:     "multipart/form-data",
								acceptPatch:    "",
							})
						}

						return
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "contract"
						// Match until "/"
						idx := strings.IndexByte(elem, '/')
						if idx < 0 {
							idx = len(elem)
						}
						args[1] = elem[:idx]
						elem = elem[idx:]

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case '/': // Prefix: "/"

							if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "name"
							// Match until "/"
							idx := strings.IndexByte(elem, '/')
							if idx < 0 {
								idx = len(elem)
							}
							args[2] = elem[:idx]
							elem = elem[idx:]

							if len(elem) == 0 {
								switch r.Method {
								case "DELETE":
									s.handleTaskDeleteRequest([3]string{
										args[0],
										args[1],
										args[2],
									}, elemIsEscaped, w, r)
								case "GET":
									s.handleTaskGetRequest([3]string{
										args[0],
										args[1],
										args[2],
									}, elemIsEscaped, w, r)
								case "PATCH":
									s.handleTaskPatchRequest([3]string{
										args[0],
										args[1],
										args[2],
									}, elemIsEscaped, w, r)
								default:
									s.notAllowed(w, r, notAllowedParams{
										allowedMethods: "DELETE,GET,PATCH",
										allowedHeaders: rn7AllowedHeaders,
										acceptPost:     "",
										acceptPatch:    "application/json",
									})
								}

								return
							}
							switch elem[0] {
							case '/': // Prefix: "/"

								if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
									elem = elem[l:]
								} else {
									break
								}

								// Param: "filename"
								// Leaf parameter, slashes are prohibited
								idx := strings.IndexByte(elem, '/')
								if idx >= 0 {
									break
								}
								args[3] = elem
								elem = ""

								if len(elem) == 0 {
									// Leaf node.
									switch r.Method {
									case "DELETE":
										s.handleTaskFileDeleteRequest([4]string{
											args[0],
											args[1],
											args[2],
											args[3],
										}, elemIsEscaped, w, r)
									case "GET":
										s.handleTaskFileGetRequest([4]string{
											args[0],
											args[1],
											args[2],
											args[3],
										}, elemIsEscaped, w, r)
									case "POST":
										s.handleTaskFilePostRequest([4]string{
											args[0],
											args[1],
											args[2],
											args[3],
										}, elemIsEscaped, w, r)
									case "PUT":
										s.handleTaskFilePutRequest([4]string{
											args[0],
											args[1],
											args[2],
											args[3],
										}, elemIsEscaped, w, r)
									default:
										s.notAllowed(w, r, notAllowedParams{
											allowedMethods: "DELETE,GET,POST,PUT",
											allowedHeaders: rn9AllowedHeaders,
											acceptPost:     "application/octet-stream",
											acceptPatch:    "",
										})
									}

									return
								}

							}

						}

					}

				}

			}

		}
	}
	s.notFound(w, r)
}

// Route is route object.
type Route struct {
	name           string
	summary        string
	operationID    string
	operationGroup string
	pathPattern    string
	count          int
	args           [4]string
}

// Name returns ogen operation name.
//
// It is guaranteed to be unique and not empty.
func (r Route) Name() string {
	return r.name
}

// Summary returns OpenAPI summary.
func (r Route) Summary() string {
	return r.summary
}

// OperationID returns OpenAPI operationId.
func (r Route) OperationID() string {
	return r.operationID
}

// OperationGroup returns the x-ogen-operation-group value.
func (r Route) OperationGroup() string {
	return r.operationGroup
}

// PathPattern returns OpenAPI path.
func (r Route) PathPattern() string {
	return r.pathPattern
}

// Args returns parsed arguments.
func (r Route) Args() []string {
	return r.args[:r.count]
}

// FindRoute finds Route for given method and path.
//
// Note: this method does not unescape path or handle reserved characters in path properly. Use FindPath instead.
func (s *Server) FindRoute(method, path string) (Route, bool) {
	return s.FindPath(method, &url.URL{Path: path})
}

// FindPath finds Route for given method and URL.
func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
	var (
		elem = u.Path
		args = r.args
	)
	if rawPath := u.RawPath; rawPath != "" {
		if normalized, ok := uri.NormalizeEscapedPath(rawPath); ok {
			elem = normalized
		}
		defer func() {
			for i, arg := range r.args[:r.count] {
				if unescaped, err := url.PathUnescape(arg); err == nil {
					r.args[i] = unescaped
				}
			}
		}()
	}

	elem, ok := s.cutPrefix(elem)
	if !ok {
		return r, false
	}

	// Static code generated router with unwrapped path search.
	switch {
	default:
		if len(elem) == 0 {
			break
		}
		switch elem[0] {
		case '/': // Prefix: "/managers"

			if l := len("/managers"); len(elem) >= l && elem[0:l] == "/managers" {
				elem = elem[l:]
			} else {
				break
			}

			if len(elem) == 0 {
				switch method {
				case "GET":
					r.name = ManagersGetOperation
					r.summary = "Получение списка менеджеров"
					r.operationID = "ManagersGet"
					r.operationGroup = "Tasks"
					r.pathPattern = "/managers"
					r.args = args
					r.count = 0
					return r, true
				default:
					return
				}
			}
			switch elem[0] {
			case '/': // Prefix: "/"

				if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
					elem = elem[l:]
				} else {
					break
				}

				// Param: "managerName"
				// Match until "/"
				idx := strings.IndexByte(elem, '/')
				if idx < 0 {
					idx = len(elem)
				}
				args[0] = elem[:idx]
				elem = elem[idx:]

				if len(elem) == 0 {
					break
				}
				switch elem[0] {
				case '/': // Prefix: "/tasks"

					if l := len("/tasks"); len(elem) >= l && elem[0:l] == "/tasks" {
						elem = elem[l:]
					} else {
						break
					}

					if len(elem) == 0 {
						switch method {
						case "GET":
							r.name = TasksGetOperation
							r.summary = "Получение списка задач"
							r.operationID = "TasksGet"
							r.operationGroup = "Tasks"
							r.pathPattern = "/managers/{managerName}/tasks"
							r.args = args
							r.count = 1
							return r, true
						case "POST":
							r.name = TaskPostOperation
							r.summary = "Создание новой задачи"
							r.operationID = "TaskPost"
							r.operationGroup = "Tasks"
							r.pathPattern = "/managers/{managerName}/tasks"
							r.args = args
							r.count = 1
							return r, true
						default:
							return
						}
					}
					switch elem[0] {
					case '/': // Prefix: "/"

						if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
							elem = elem[l:]
						} else {
							break
						}

						// Param: "contract"
						// Match until "/"
						idx := strings.IndexByte(elem, '/')
						if idx < 0 {
							idx = len(elem)
						}
						args[1] = elem[:idx]
						elem = elem[idx:]

						if len(elem) == 0 {
							break
						}
						switch elem[0] {
						case '/': // Prefix: "/"

							if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
								elem = elem[l:]
							} else {
								break
							}

							// Param: "name"
							// Match until "/"
							idx := strings.IndexByte(elem, '/')
							if idx < 0 {
								idx = len(elem)
							}
							args[2] = elem[:idx]
							elem = elem[idx:]

							if len(elem) == 0 {
								switch method {
								case "DELETE":
									r.name = TaskDeleteOperation
									r.summary = "Удалить задачу"
									r.operationID = "TaskDelete"
									r.operationGroup = "Tasks"
									r.pathPattern = "/managers/{managerName}/tasks/{contract}/{name}"
									r.args = args
									r.count = 3
									return r, true
								case "GET":
									r.name = TaskGetOperation
									r.summary = "Получение задачи"
									r.operationID = "TaskGet"
									r.operationGroup = "Tasks"
									r.pathPattern = "/managers/{managerName}/tasks/{contract}/{name}"
									r.args = args
									r.count = 3
									return r, true
								case "PATCH":
									r.name = TaskPatchOperation
									r.summary = "Частичное обновление задачи"
									r.operationID = "TaskPatch"
									r.operationGroup = "Tasks"
									r.pathPattern = "/managers/{managerName}/tasks/{contract}/{name}"
									r.args = args
									r.count = 3
									return r, true
								default:
									return
								}
							}
							switch elem[0] {
							case '/': // Prefix: "/"

								if l := len("/"); len(elem) >= l && elem[0:l] == "/" {
									elem = elem[l:]
								} else {
									break
								}

								// Param: "filename"
								// Leaf parameter, slashes are prohibited
								idx := strings.IndexByte(elem, '/')
								if idx >= 0 {
									break
								}
								args[3] = elem
								elem = ""

								if len(elem) == 0 {
									// Leaf node.
									switch method {
									case "DELETE":
										r.name = TaskFileDeleteOperation
										r.summary = "Удалить файл"
										r.operationID = "TaskFileDelete"
										r.operationGroup = "Tasks"
										r.pathPattern = "/managers/{managerName}/tasks/{contract}/{name}/{filename}"
										r.args = args
										r.count = 4
										return r, true
									case "GET":
										r.name = TaskFileGetOperation
										r.summary = "Получение файла"
										r.operationID = "TaskFileGet"
										r.operationGroup = "Tasks"
										r.pathPattern = "/managers/{managerName}/tasks/{contract}/{name}/{filename}"
										r.args = args
										r.count = 4
										return r, true
									case "POST":
										r.name = TaskFilePostOperation
										r.summary = "Загрузка файла"
										r.operationID = "TaskFilePost"
										r.operationGroup = "Tasks"
										r.pathPattern = "/managers/{managerName}/tasks/{contract}/{name}/{filename}"
										r.args = args
										r.count = 4
										return r, true
									case "PUT":
										r.name = TaskFilePutOperation
										r.summary = "Загрузка файла с заменой"
										r.operationID = "TaskFilePut"
										r.operationGroup = "Tasks"
										r.pathPattern = "/managers/{managerName}/tasks/{contract}/{name}/{filename}"
										r.args = args
										r.count = 4
										return r, true
									default:
										return
									}
								}

							}

						}

					}

				}

			}

		}
	}
	return r, false
}